Bifurcate `doc-client`, splitting out `doc-ui`.
The former doesn't have DOM-ish stuff, and the latter does.
