+++ local-modules/@bayou/data-model-client/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
     "@bayou/deps-react": "local",
-    "@bayou/doc-client": "local"
+    "@bayou/doc-ui": "local"
   }
 }
