+++ local-modules/@bayou/doc-ui/package.json
@@ -1,0 +1,16 @@
+{
+  "dependencies": {
+    "@bayou/api-common": "local",
+    "@bayou/config-client": "local",
+    "@bayou/data-model-client": "local",
+    "@bayou/doc-client": "local",
+    "@bayou/doc-common": "local",
+    "@bayou/promise-util": "local",
+    "@bayou/quill-util": "local",
+    "@bayou/see-all": "local",
+    "@bayou/typecheck": "local",
+    "@bayou/ui-components": "local",
+    "@bayou/util-client": "local",
+    "@bayou/util-common": "local"
+  }
+}
