+++ local-modules/@bayou/top-client/package.json
@@ -6,7 +6,7 @@
     "@bayou/api-common": "local",
     "@bayou/app-common": "local",
     "@bayou/config-client": "local",
-    "@bayou/doc-client": "local",
+    "@bayou/doc-ui": "local",
     "@bayou/env-client": "local",
     "@bayou/see-all": "local",
     "@bayou/see-all-client": "local",
