+++ local-modules/@bayou/ui-components/package.json
@@ -1,7 +1,7 @@
 {
   "dependencies": {
     "@bayou/deps-react": "local",
-    "@bayou/doc-client": "local",
+    "@bayou/doc-ui": "local",
     "@bayou/ui-common": "local"
   }
 }
